# Add batcher timestamp to report queue time properly and to maintain batcher functionality