--- a/src/core/infer_request.cc
+++ b/src/core/infer_request.cc
@@ -1,4 +1,4 @@
-// Copyright 2020-2021, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2020-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -559,6 +559,7 @@ InferenceRequest::PrepareForInference()
 
   // Clear the timestamps
   queue_start_ns_ = 0;
+  batcher_start_ns_ = 0;
 #ifdef TRITON_ENABLE_STATS
   request_start_ns_ = 0;
 #endif  // TRITON_ENABLE_STATS