--- a/src/core/instance_queue.cc
+++ b/src/core/instance_queue.cc
@@ -1,4 +1,4 @@
-// Copyright 2021, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2021-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -75,7 +75,7 @@ InstanceQueue::Dequeue(
         size_t batch_size = (*payload)->BatchSize();
         if ((!payload_queue_.empty()) &&
             (!payload_queue_.front()->IsSaturated()) &&
-            (now_ns - payload_queue_.front()->QueueStartNs()) >
+            (now_ns - payload_queue_.front()->BatcherStartNs()) >
                 max_queue_delay_ns_) {
           std::lock_guard<std::mutex> exec_lock(
               *(payload_queue_.front()->GetExecMutex()));