--- a/src/core/payload.cc
+++ b/src/core/payload.cc
@@ -1,4 +1,4 @@
-// Copyright 2021, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright 2021-2022, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -32,7 +32,7 @@ Payload::Payload()
     : op_type_(Operation::INFER_RUN),
       requests_(std::vector<std::unique_ptr<InferenceRequest>>()),
       OnCallback_([]() {}), OnSecondaryCallback_([]() {}), instance_(nullptr),
-      state_(State::UNINITIALIZED), queue_start_ns_(0), saturated_(true)
+      state_(State::UNINITIALIZED), batcher_start_ns_(0), saturated_(true)
 {
   exec_mu_.reset(new std::mutex());
 }
@@ -77,7 +77,7 @@ Payload::Reset(const Operation op_type, TritonModelInstance* instance)
   instance_ = instance;
   state_ = State::UNINITIALIZED;
   status_.reset(new std::promise<Status>());
-  queue_start_ns_ = 0;
+  batcher_start_ns_ = 0;
 }
 
 void
@@ -89,7 +89,7 @@ Payload::Release()
   OnSecondaryCallback_ = []() {};
   instance_ = nullptr;
   state_ = State::RELEASED;
-  queue_start_ns_ = 0;
+  batcher_start_ns_ = 0;
 }
 
 size_t
@@ -111,8 +111,9 @@ Payload::ReserveRequests(size_t size)
 void
 Payload::AddRequest(std::unique_ptr<InferenceRequest> request)
 {
-  if ((queue_start_ns_ == 0) || (queue_start_ns_ > request->QueueStartNs())) {
-    queue_start_ns_ = request->QueueStartNs();
+  if ((batcher_start_ns_ == 0) ||
+      (batcher_start_ns_ > request->BatcherStartNs())) {
+    batcher_start_ns_ = request->BatcherStartNs();
   }
   requests_.push_back(std::move(request));
 }