Now in the case of oldest sequence batcher, the queue time is tracked (not counting towards overhead)
Before:
```
...
Avg request latency: 68538 usec (overhead 60622 usec + queue 2080 usec + compute input 172 usec + compute infer 5525 usec + compute output 139 usec)
...
```
After:
```
...
Avg request latency: 69777 usec (overhead 168 usec + queue 63749 usec + compute input 168 usec + compute infer 5554 usec + compute output 138 usec)
...
```